requirements.txt
fastapi~=0.115.6
python-multipart~=0.0.20
starlette~=0.41.3
motor~=3.6.0
pymongo~=4.9.2
python-dotenv~=1.0.1